requirements.txt
pyhamcrest
pytest
pytest_watch
lxml
cairosvg